+++ v1/ast/syncpools.go
@@ -2,7 +2,6 @@ package ast
 
 import (
 	"bytes"
-	"strings"
 	"sync"
 
 	"github.com/open-policy-agent/opa/v1/util"
@@ -12,15 +11,7 @@ var (
 	TermPtrPool     = util.NewSyncPool[Term]()
 	BytesReaderPool = util.NewSyncPool[bytes.Reader]()
 	IndexResultPool = util.NewSyncPool[IndexResult]()
-	bbPool          = util.NewSyncPool[bytes.Buffer]()
-	// Needs custom pool because of custom Put logic.
-	sbPool = &stringBuilderPool{
-		pool: sync.Pool{
-			New: func() any {
-				return &strings.Builder{}
-			},
-		},
-	}
+
 	// Needs custom pool because of custom Put logic.
 	varVisitorPool = &vvPool{
 		pool: sync.Pool{
@@ -31,18 +22,8 @@ var (
 	}
 )
 
-type (
-	stringBuilderPool struct{ pool sync.Pool }
-	vvPool            struct{ pool sync.Pool }
-)
-
-func (p *stringBuilderPool) Get() *strings.Builder {
-	return p.pool.Get().(*strings.Builder)
-}
-
-func (p *stringBuilderPool) Put(sb *strings.Builder) {
-	sb.Reset()
-	p.pool.Put(sb)
+type vvPool struct {
+	pool sync.Pool
 }
 
 func (p *vvPool) Get() *VarVisitor {